* Recalculate all surrounding road tiles, including corners
This was not necessary before, but now the corners also need to be
retiled because they might need to merge with other road cells.

The actual change in this commit happened to world.building.path's
recalculate_surrounding_tile_orientation which is not included in
this commit because the default parameter was changed to match the
one of get_surrounding() in shapes.rect.